+++ app/src/components/layout/HomeContainer.tsx
@@ -1,15 +1,5 @@
 import React from "react";
 
-export default function HomeContainer({ title, img, children }: { title: React.ReactNode; img: string; children: React.ReactNode }) {
-  return (
-    <div className="max-w-7xl mx-auto bg-white flex flex-col xl:flex-row-reverse xl:justify-between gap-2 p-[1rem] md:p-[4rem]">
-      <div className="w-64 md:w-96 mx-auto flex-none md:hidden xl:block">
-        <img src={img} width="auto" height="auto" className="md:px-0 mx-auto" />
-      </div>
-      <div className="max-w-2xl">
-        <h1 className="font-bold text-3xl md:text-5xl leading-10 md:leading-tight mt-2">{title}</h1>
-        {children}
-      </div>
-    </div>
-  );
+export default function HomeContainer({ children }: { children: React.ReactNode }) {
+  return <div className="max-w-7xl mx-auto p-[1rem] md:p-[4rem]">{children}</div>;
 }
