+++ app/src/scenes/home/components/Engagement.tsx
@@ -10,16 +10,16 @@ export default function Engagement() {
   if (isPending) return <Loader />;
   if (isError) return <div>Erreur</div>;
   return (
-    <section id="engagements">
+    <section id="engagements" className="max-w-6xl mx-auto p-[1rem] md:p-[0rem]">
       <p className="font-bold leading-loose">Découvrez d’autres formes d’engagement</p>
       <p className="mt-1 text-gray-500">Si l’engagement vous donne envie, vous trouverez des dispositifs qui pourront vous intéresser.</p>
-      <div className="flex gap-6 overflow-x-auto mt-3 pb-2">
+      <div className="flex gap-6 overflow-x-auto mt-3 pb-3">
         {programs.map((program) => {
           return <EngagementCard key={program._id} program={program} />;
         })}
       </div>
-      <Link className="w-full" to="/public-engagements">
-        <div className="mt-2 text-sm border-[1px] border-gray-300 px-3 py-2 text-center rounded-md">Découvrir les possibilités d'engagement</div>
+      <Link to="/public-engagements">
+        <div className="w-full md:w-fit mt-1 text-sm border-[1px] border-gray-300 px-6 py-2 text-center rounded-md">Découvrir les possibilités d'engagement</div>
       </Link>
     </section>
   );
